GEOMESA-740: Improve GeoMesa Tools Configure Behavior
This small fix causes the configuration script for geo tools
to correctly prompt the user if the home or libs variables
are already set. Without this, it's possible to end up
incorrectly configuring the paths as a result of prior runs.

This is unlikely to affect users, but could lead to weirdness
in certain specific development situations.

Signed-off-by: Jake Kurzer <jake.kurzer@ccri.com>

updating to accumulo 1.6

updates

Quick